establishing the update_photo route
Almost done with the remaining piece being how to route action
update_photo so that when hitting the "Submit" button on edit_form, the
process is completed. As is, the original information [caption, photos]
remains in the input fields after hitting "Submit" with the page
essentially refreshing to include the same information. URL does update
however.